Buffer size greatly increased and results trimmed. Argument now taken for file to filter.
Filter.java- message buffer increased from 80 to 900, arbitrarily
chosen. There's the principle of emulating KHUx's messaging system, and
then there's just having a actually useful program- and its time for the
latter. In the case of running with FilterTest, this limit determines
the max number of characters each line can be. Excess null characters
which were previously carried over due to the buffer size exceeding the
message length (which is...always) are now trimmed off.

FilterTest.java- is no longer hardcoded to take "ass.txt" as a test
file, and instead now requires a single argument to determine which file
to filter.  The case to check a single line of user input (fully
functional) has been commented out.

As a result of all of the changes above, redirecting FilterTest's output
via '>' now cleanly produces a copy of the test file, fully filtered.